# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkyMP Backend API
  version: 2.0.0
  description: Canonical API v2 for launchers, dashboards, and SkyMP game servers.
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v2/admin/players:
    get:
      tags: [Administration]
      operationId: listPlayers
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/Collection' }
        '401': { $ref: '#/components/responses/JsonError' }
    post:
      tags: [Administration]
      operationId: createPlayer
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '201': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/admin/players/{profileId}:
    parameters:
      - $ref: '#/components/parameters/ProfileId'
    get:
      tags: [Administration]
      operationId: getPlayer
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
    put:
      tags: [Administration]
      operationId: updatePlayer
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/admin/players/{profileId}/whitelist:
    put:
      tags: [Administration]
      operationId: updatePlayerWhitelist
      security: [{ dashboardBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ProfileId' }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/admin/players/{profileId}/ban:
    put:
      tags: [Administration]
      operationId: updatePlayerBan
      security: [{ dashboardBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ProfileId' }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
components:
  securitySchemes:
    dashboardBearer:
      type: http
      scheme: bearer
  parameters:
...
    ProfileId:
      name: profileId
      in: path
      required: true
      schema: { type: integer }
...
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    JsonSuccess:
      description: Successful JSON response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Collection:
      description: Collection response
      content:
        application/json:
          schema:
            type: object
            required: [items, total]
            properties:
              items: { type: array, items: {} }
              total: { type: integer, minimum: 0 }
    JsonError:
      description: Structured API v2 error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}